requirements.txt
Django
djangorestframework
web3
bitcoinaddress